Merge team-3-damagebuff into main

+++ source/core/src/main/com/csse3200/game/components/player/PlayerStatsDisplay.java
@@ -261,7 +261,7 @@ public class PlayerStatsDisplay extends UIComponent {
         speedProgressBar.setValue(speedPercentage);
     }
 
-    public void updateDamageUI(float damage) {
+    public void updateDamageUI(int damage) {
         damageProgressBar.setValue(damage);
     }
